requirements.txt
PyQt6>=6.4.0
python-vlc>=3.0.0
aiohttp>=3.8.0
pillow>=9.0.0
cairosvg>=2.5.2